% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Function to run dijkstra algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{as data.frame}
}
\value{
a vector with smallest length from the initial node to all the nodes
}
\description{
Function to run dijkstra algorithm
}
