% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Implements Euclidean algorithm to find GCD in the fastest possible way
Authors@R: Aashana, Jasleen}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}
}
\value{
GCD for x and y
}
\description{
Implements Euclidean algorithm to find GCD in the fastest possible way
Authors@R: Aashana, Jasleen
}
\references{
<https://en.wikipedia.org/wiki/Euclidean_algorithm>
}
