% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki_graph dataset to be used as an graph argument for dijkstra function
data.frame with three variables (v1, v2 and w)
v1 and v2 are the edges of the graph with the weight of the edge w}
\format{An object of class \code{data.frame} with 18 rows and 3 columns.}
\usage{
data(wiki_graph)
}
\description{
wiki_graph dataset to be used as an graph argument for dijkstra function
data.frame with three variables (v1, v2 and w)
v1 and v2 are the edges of the graph with the weight of the edge w
}
\references{
<https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm>
}
\keyword{datasets}
