% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Function to run dijkstra algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{as data.frame}

\item{init_node}{as numeric}
}
\value{
a vector with smallest length from the initial node to all the nodes
}
\description{
Function to run dijkstra algorithm
}
\examples{
dijkstra(data.frame(v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6), v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9)), 1)
}
